feat: add support for defining an --include file
My team has a use case for always requiring certain files to be included in the package.xml (these include a generated class file and test). And these files change based off of values (i.e. CI build numbers, etc) with each job run. We don't want to be submitting a new commit with each job build, but we need to include these files in the package.xml for `source:deploy`. Using this solution, we can ensure that the generated files are always included (along with `git add -f`).